Add files via upload
This is a program that assigns seats on an airplane.  The airplane has 10 seats.  The program asks if the passenger requires a vegetarian meal or a standard meal.  There are only 5 vegetarian meals available.  The seats are numbered 1-10 and no double seats are assigned.  If a passenger requires a vegetarian meal but it is not available , the system gives a choice of standard meal or printout "Next flight with a vegetarian option leaves in 4 hours".  The final printout is the First Name , Last Name (of the all the passengers), seat number, and a V for Vegetarian or S for standard meal. If all the seats are taken, the program tells the user so and the time of next flight.